fix: 프로필세팅, 셀렉 수정 %155

+++ src/components/profileSettings/ProfileSettingsComponent.js
@@ -42,8 +42,8 @@ const ProfileSettingsComponent = ({ history }) => {
     statusMessage: '',
     isOpen: false,
   });
-  const [nicknameLength, setNicknameLength] = useState(false);
   const [nicknameExists, setNicknameExists] = useState(false);
+  const [nicknameRoles, setNicknameRoles] = useState(false);
   const [statusMessageOverCount, setStatusMessageOverCount] = useState(false);
   const viewSize = useResponsive();
   const [nickname, setNickname] = useState(sessionStorage.getItem('nickname'));
@@ -61,7 +61,7 @@ const ProfileSettingsComponent = ({ history }) => {
       nickname: e.target.value,
       changed: true,
     });
-    setNicknameLength(false);
+    setNicknameRoles(false);
   };
 
   const onStatusMessageChange = (e) => {
@@ -107,26 +107,38 @@ const ProfileSettingsComponent = ({ history }) => {
   };
 
   const onUpdataSubmitHandler = async () => {
-    if (nicknameLength > 2) {
-      const existsResponse = await client.get('/api/v1/member/exists', { params: { nickname: myInfo.nickname } });
-      existsResponse.data.data === true
-        ? myInfo.nickname === nickname
-          ? setNicknameExists(false)
-          : setNicknameExists(true)
-        : setNicknameExists(false);
-      myInfo.statusMessage.length < 30 ? setStatusMessageOverCount(false) : setStatusMessageOverCount(true);
-      if ((existsResponse.data.data === false || myInfo.nickname === nickname) && myInfo.statusMessage.length < 30) {
-        const upDataResponse = await client.patch('/api/v1/member/me', myInfo);
-        if (upDataResponse.data.message === 'update') {
-          alert('성공적으로 변경됐습니다 :)');
-          setNickname(myInfo.nickname);
-          sessionStorage.setItem('nickname', myInfo.nickname);
-          localStorage.setItem('nickname', myInfo.nickname);
-          window.location.replace(`/${myInfo.nickname}/settings`);
-        }
-      }
+    // 닉네임이 있나?
+    const existsResponse = await client.get('/api/v1/member/exists', { params: { nickname: myInfo.nickname } });
+    existsResponse.data.data === true
+      ? myInfo.nickname === nickname
+        ? setNicknameExists(false)
+        : setNicknameExists(true)
+      : setNicknameExists(false);
+    //상테메시지 30줄보다 짧은가?
+    myInfo.statusMessage.length < 30 ? setStatusMessageOverCount(false) : setStatusMessageOverCount(true);
+    // 조건에 맞나?
+    const regExp = /^[0-9a-zA-Z]([_]?[0-9a-zA-Z]){2,19}$/; // 영문, 숫자, 특수문자 '_' 를 포함한 3~20자 특수문자는 마지막에 못옴
+    if (!regExp.test(myInfo.nickname)) {
+      // 닉네임 조건에 부합하지 않음
+      setNicknameRoles(true);
     } else {
-      setNicknameLength(true);
+      setNicknameRoles(false);
+    }
+
+    // 모든 조건을 만족하는가? (위의 3조건 + 지금 닉네임과 동일한가)
+    if (
+      (existsResponse.data.data === false || myInfo.nickname === nickname) &&
+      myInfo.statusMessage.length < 30 &&
+      regExp.test(myInfo.nickname)
+    ) {
+      const upDataResponse = await client.patch('/api/v1/member/me', myInfo);
+      if (upDataResponse.data.message === 'update') {
+        alert('성공적으로 변경됐습니다 :)');
+        setNickname(myInfo.nickname);
+        sessionStorage.setItem('nickname', myInfo.nickname);
+        localStorage.setItem('nickname', myInfo.nickname);
+        window.location.replace(`/${myInfo.nickname}/settings`);
+      }
     }
   };
 
@@ -155,7 +167,7 @@ const ProfileSettingsComponent = ({ history }) => {
       <MainWrapper>
         <ProfileImg
           type="file"
-          accept="image/x-png,image/gif,image/jpeg"
+          accept="image/x-png, image/jpeg, image/jpg"
           onChange={onFileChange}
           style={{
             backgroundImage: `url(${myInfo.imgUrl})`,
@@ -176,9 +188,13 @@ const ProfileSettingsComponent = ({ history }) => {
           <InputBoxWrapper>
             <EachTitle>별명</EachTitle>
             <InputBox placeholder={myInfo.nickname} value={myInfo.nickname} onChange={onNicknameChange} />
-            {nicknameLength ? <AlertMessage>앗, 별명이 너무 짧아요! 3자 이상 입력해주세요.</AlertMessage> : <></>}
             {nicknameExists ? (
               <AlertMessage>앗, 누군가 이미 사용중인 별명이네요. 다른 별명을 사용해보세요.</AlertMessage>
+            ) : (
+              <></>
+            )}
+            {nicknameRoles ? (
+              <AlertMessage>앗, 숫자/영문/_만 사용하여 3-20자 이내로 사용해보세요.</AlertMessage>
             ) : (
               <></>
             )}
